% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\alias{as.sand}
\alias{is.sand}
\alias{sdesc}
\alias{sdesc<-}
\alias{smeta}
\alias{smeta<-}
\title{SAND access and assignment functions}
\usage{
is.sand(x)

as.sand(x)

smeta(x)

smeta(x) <- value

sdesc(x)

sdesc(x) <- value
}
\arguments{
\item{x}{anything}

\item{value}{object on right side of assignment}
}
\description{
SAND access and assignment functions
}

