\name{aidsblog}
\alias{aidsblog}
\title{AIDS blog citation network}
\description{
  A snapshot of the pattern of citation among 146 unique blogs 
  related to AIDS, patients, and their support networks,
  collected by Suchi Gopal (see reference below) over a randomly
  selected three-day period in August 2005. A directed edge from one
  blog to another indicates that the former has a link to the latter in
  their web page (more specifically, the former refers to the latter in
  their so-called `blogroll').
}
\usage{
aidsblog
}
\format{
  A directed igraph graph object with 146 vertices and 187 edges.
}
\source{
  This dataset was provided to us by Suchi Gopal.
  Please cite the reference below if you use this dataset in your work.
}
\references{
  S. Gopal, The evolving social geography of blogs. In Societies and
  Cities in the Age of Instant Access, ed. by H. Miller (Springer,
  Berlin, 2007), 139 pp. 275-294.
}
\keyword{datasets}
