% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqldb.R
\name{open_sand_db}
\alias{open_sand_db}
\title{Open a SAND SQLite database}
\usage{
open_sand_db(dbname)
}
\arguments{
\item{dbname}{name of the sqlite database}
}
\value{
SQLiteConnection
}
\description{
Open a SAND SQLite database
}

