% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parsers}
\alias{parsers}
\alias{format_types}
\alias{unformat_types}
\title{Type parsers for sander}
\usage{
format_types(x)

unformat_types(x)
}
\arguments{
\item{x}{a string}
}
\description{
Type parsers for sander
}
