% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_sand}
\alias{read_sand}
\title{Read an unnested SAND directory}
\usage{
read_sand(x, data_has_header = TRUE, meta_has_header = TRUE,
  type_has_header = TRUE, col_types = NULL, rdata = read_data,
  rdesc = read_desc, rmeta = read_meta, rtype = read_type)
}
\arguments{
\item{x}{directory name}

\item{data_has_header}{does the data file have headers?}

\item{meta_has_header}{does the COLUMN file have headers?}

\item{type_has_header}{does the TYPE file have headers?}

\item{col_types}{data types for each column (see readr documentation)}

\item{rdata}{main data file reader}

\item{rdesc}{description file reader}

\item{rmeta}{metadata file reader}

\item{rtype}{type file reader}
}
\value{
a sand object
}
\description{
Read an unnested SAND directory
}

