% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_sand}
\alias{write_sand}
\title{Write SAND object to SAND formatted directory}
\usage{
write_sand(x, path)
}
\arguments{
\item{x}{a sand object}

\item{path}{name of directory that will be created}
}
\description{
Write SAND object to SAND formatted directory
}
\seealso{
Other io: \code{\link{open_sand_db}},
  \code{\link{read_sand}}, \code{\link{write_to_db}}
}

