% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqldb.R
\name{write_to_db}
\alias{write_to_db}
\title{Write SAND object to SQLite database}
\usage{
write_to_db(x, tblname = deparse(substitute(x)), dbname = ":memory:")
}
\arguments{
\item{x}{SAND object}

\item{tblname}{name of new data table that will be created}

\item{dbname}{name of SQLite database}
}
\value{
SQLiteConnection
}
\description{
Write SAND object to SQLite database
}

