% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\alias{is.sand}
\alias{as.sand}
\alias{desc}
\alias{meta}
\alias{type}
\alias{desc<-}
\alias{meta<-}
\alias{type<-}
\alias{field_info}
\title{SAND access and assignment functions}
\usage{
is.sand(x)

as.sand(x)

desc(x)

meta(x)

type(x)

desc(x) <- value

meta(x) <- value

type(x) <- value

field_info(x, field)
}
\arguments{
\item{x}{anything}

\item{value}{object on right side of assignment}

\item{field}{a column name from the dataset}
}
\description{
Functions for accessing elements of a SAND object, accessing and casting
sand classes, and viewing documentation.
}
\details{
\code{as.sand} is particularly noteworthy. Any data frame can be case as a
sand object. Then it can be written into a SAND format data directory with
the \code{write_sand} function. This will automatically generate default
COLUMN.tsv and TYPE.tsv files. So this is the starting point for building a
new SAND project.
}
