% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\alias{as.sand}
\alias{desc}
\alias{desc<-}
\alias{field_info}
\alias{is.sand}
\alias{meta}
\alias{meta<-}
\title{SAND access and assignment functions}
\usage{
is.sand(x)

as.sand(x)

meta(x)

field_info(x, field)

meta(x) <- value

desc(x)

desc(x) <- value
}
\arguments{
\item{x}{anything}

\item{field}{a column name from the dataset}

\item{value}{object on right side of assignment}
}
\description{
SAND access and assignment functions
}

