% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_sand}
\alias{read_sand}
\title{Read an unnested SAND directory}
\usage{
read_sand(x, rdata = read_data, rdesc = read_desc, rmeta = read_meta)
}
\arguments{
\item{x}{directory name}

\item{rdata}{main data file reader}

\item{rdesc}{description file reader}

\item{rmeta}{metadata file reader}
}
\value{
a sand object
}
\description{
Read an unnested SAND directory
}

