% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_sand}
\alias{read_sand}
\title{Read an unnested SAND directory}
\usage{
read_sand(x = getwd(), data_has_header = TRUE, meta_has_header = TRUE,
  type_has_header = TRUE, col_types = NULL, rdata = read_data,
  rdesc = read_desc, rmeta = read_meta, rtype = read_type)
}
\arguments{
\item{x}{directory name, defaults to working directory}

\item{data_has_header}{whether data file has headers}

\item{meta_has_header}{whether COLUMN file has headers}

\item{type_has_header}{whether TYPE file has headers}

\item{col_types}{data types for each column (see details)}

\item{rdata}{main data file reader}

\item{rdesc}{description file reader}

\item{rmeta}{metadata file reader}

\item{rtype}{type file reader}
}
\value{
a sand object
}
\description{
Read an unnested SAND directory
}
\examples{
diamonds_dir <- system.file('extdata', 'diamonds', package='sandr')
x <- read_sand(diamonds_dir)
}
\seealso{
Other io: \code{\link{open_sand_db}},
  \code{\link{write_sand}}, \code{\link{write_to_db}}
}
